add createUser + refactor

+++ src/main/kotlin/com/ulu/fetchers/AttributeCategoryFetcher.kt
@@ -10,6 +10,7 @@ import jakarta.inject.Singleton
 @Singleton
 class AttributeCategoryFetcher(
     private val attributeCategoryRepository: AttributeCategoryRepository,
+    private val requestValidatorService: RequestValidatorService,
     private val securityService: DefaultSecurityService,
 ) {
     fun getAttributeCategories(): DataFetcher<List<AttributeCategory>> {
@@ -20,7 +21,7 @@ class AttributeCategoryFetcher(
 
     fun createAttributeCategory(): DataFetcher<AttributeCategory> {
         return DataFetcher {
-            RequestValidatorService().verifyAdmin(securityService)
+            requestValidatorService.verifyAdmin(securityService)
             val name: String = it.getArgument("name")
             if (attributeCategoryRepository.existsByName(name)) {
                 error("AttributeCategory with identical name already exists")
@@ -31,7 +32,7 @@ class AttributeCategoryFetcher(
 
     fun editAttributeCategory(): DataFetcher<AttributeCategory> {
         return DataFetcher {
-            RequestValidatorService().verifyAdmin(securityService)
+            requestValidatorService.verifyAdmin(securityService)
 
             val attributeCategoryId = it.getArgument<String>("id").toLong()
             val name: String = it.getArgument("name")
@@ -50,7 +51,7 @@ class AttributeCategoryFetcher(
 
     fun deleteAttributeCategory(): DataFetcher<String> {
         return DataFetcher {
-            RequestValidatorService().verifyAdmin(securityService)
+            requestValidatorService.verifyAdmin(securityService)
 
             attributeCategoryRepository.deleteById(it.getArgument("id"))
             return@DataFetcher "deleted"
